delete confirm

+++ src/scripts/list.js
@@ -67,9 +67,12 @@ function createListItem(item) {
   deleteButton.style.display = "none";
 
   deleteButton.onclick = (event) => {
+    if(window.confirm(`Are you sure you would like to delete the "${item.title}"?`)) {
+      event.stopPropagation();
+      listItem.remove();
+      deleteJournal(timestamp);
+    }
     event.stopPropagation();
-    listItem.remove();
-    deleteJournal(timestamp);
   };
 
   listItem.appendChild(deleteButton);
